# 8349727: [PPC] C1: Improve Class.isInstance intrinsic