PPC64 implementation of [JDK-8337251](https://bugs.openjdk.org/browse/JDK-8337251).
The new runtime stub is called like a C function. The initial version therefore used a `FunctionDescriptor` with relocation on PPC64 with ABIv1. I've changed that with the 3rd Commit. `rt_call` jumps directly to the entry point, now. 

Performance measured on Power10: `make run-test TEST="micro:SecondarySupersLookup" MICRO="VM_OPTIONS=-XX:TieredStopAtLevel=1"`

Before this patch (C code)
```
Benchmark                             Mode  Cnt   Score   Error  Units
SecondarySupersLookup.testNegative00  avgt   15  18.570 ± 0.009  ns/op
...
SecondarySupersLookup.testNegative30  avgt   15  18.566 ± 0.002  ns/op
SecondarySupersLookup.testNegative32  avgt   15  19.177 ± 1.347  ns/op
SecondarySupersLookup.testNegative40  avgt   15  18.569 ± 0.006  ns/op
SecondarySupersLookup.testNegative50  avgt   15  19.207 ± 1.334  ns/op
SecondarySupersLookup.testNegative55  avgt   15  19.708 ± 1.338  ns/op
SecondarySupersLookup.testNegative56  avgt   15  19.132 ± 0.137  ns/op
SecondarySupersLookup.testNegative57  avgt   15  19.133 ± 0.134  ns/op
SecondarySupersLookup.testNegative58  avgt   15  19.772 ± 1.316  ns/op
SecondarySupersLookup.testNegative59  avgt   15  19.109 ± 0.014  ns/op
SecondarySupersLookup.testNegative60  avgt   15  22.381 ± 0.016  ns/op
SecondarySupersLookup.testNegative61  avgt   15  22.331 ± 0.011  ns/op
SecondarySupersLookup.testNegative62  avgt   15  22.352 ± 0.029  ns/op
SecondarySupersLookup.testNegative63  avgt   15  30.371 ± 0.031  ns/op
SecondarySupersLookup.testNegative64  avgt   15  29.927 ± 0.221  ns/op
SecondarySupersLookup.testPositive01  avgt   15  18.571 ± 0.006  ns/op
...
SecondarySupersLookup.testPositive09  avgt   15  18.599 ± 0.140  ns/op
SecondarySupersLookup.testPositive10  avgt   15  19.210 ± 1.332  ns/op
SecondarySupersLookup.testPositive16  avgt   15  18.603 ± 0.142  ns/op
SecondarySupersLookup.testPositive20  avgt   15  19.210 ± 1.333  ns/op
SecondarySupersLookup.testPositive30  avgt   15  18.600 ± 0.140  ns/op
SecondarySupersLookup.testPositive32  avgt   15  18.637 ± 0.189  ns/op
SecondarySupersLookup.testPositive40  avgt   15  19.137 ± 0.190  ns/op
SecondarySupersLookup.testPositive50  avgt   15  18.567 ± 0.002  ns/op
SecondarySupersLookup.testPositive60  avgt   15  19.069 ± 0.004  ns/op
SecondarySupersLookup.testPositive63  avgt   15  26.024 ± 0.017  ns/op
SecondarySupersLookup.testPositive64  avgt   15  29.932 ± 1.002  ns/op
```

After this patch (assembler code):
```
Benchmark                             Mode  Cnt   Score   Error  Units
SecondarySupersLookup.testNegative00  avgt   15  18.889 ± 0.016  ns/op
...
SecondarySupersLookup.testNegative30  avgt   15  18.830 ± 0.222  ns/op
SecondarySupersLookup.testNegative32  avgt   15  18.885 ± 0.006  ns/op
SecondarySupersLookup.testNegative40  avgt   15  18.884 ± 0.003  ns/op
SecondarySupersLookup.testNegative50  avgt   15  18.885 ± 0.004  ns/op
SecondarySupersLookup.testNegative55  avgt   15  18.775 ± 0.287  ns/op
SecondarySupersLookup.testNegative56  avgt   15  18.877 ± 0.008  ns/op
SecondarySupersLookup.testNegative57  avgt   15  18.821 ± 0.211  ns/op
SecondarySupersLookup.testNegative58  avgt   15  18.874 ± 0.006  ns/op
SecondarySupersLookup.testNegative59  avgt   15  18.991 ± 0.344  ns/op
SecondarySupersLookup.testNegative60  avgt   15  18.876 ± 0.195  ns/op
SecondarySupersLookup.testNegative61  avgt   15  18.827 ± 0.265  ns/op
SecondarySupersLookup.testNegative62  avgt   15  19.009 ± 0.182  ns/op
SecondarySupersLookup.testNegative63  avgt   15  61.667 ± 0.341  ns/op
SecondarySupersLookup.testNegative64  avgt   15  62.559 ± 0.119  ns/op
SecondarySupersLookup.testPositive01  avgt   15  18.766 ± 0.295  ns/op
...
SecondarySupersLookup.testPositive09  avgt   15  18.872 ± 0.007  ns/op
SecondarySupersLookup.testPositive10  avgt   15  18.869 ± 0.005  ns/op
SecondarySupersLookup.testPositive16  avgt   15  18.869 ± 0.004  ns/op
SecondarySupersLookup.testPositive20  avgt   15  18.869 ± 0.006  ns/op
SecondarySupersLookup.testPositive30  avgt   15  18.921 ± 0.260  ns/op
SecondarySupersLookup.testPositive32  avgt   15  18.766 ± 0.296  ns/op
SecondarySupersLookup.testPositive40  avgt   15  18.763 ± 0.189  ns/op
SecondarySupersLookup.testPositive50  avgt   15  18.869 ± 0.008  ns/op
SecondarySupersLookup.testPositive60  avgt   15  18.718 ± 0.258  ns/op
SecondarySupersLookup.testPositive63  avgt   15  34.618 ± 0.037  ns/op
SecondarySupersLookup.testPositive64  avgt   15  52.376 ± 0.463  ns/op
```

The patch avoids the slowdown until 62. The 63 and 64 cases were slow due to https://github.com/openjdk/jdk/blob/b6443f6ff96707f67552df41c01d18c193560223/src/hotspot/cpu/ppc/macroAssembler_ppc.cpp#L2556 which is assumed to be not so relevant in real world applications?

4th Commit unrolls the `repne_scan` loop to avoid the performance regression:
```
SecondarySupersLookup.testPositive63  avgt   15  24.314 ± 0.102  ns/op
SecondarySupersLookup.testPositive64  avgt   15  28.907 ± 0.045  ns/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8349727](https://bugs.openjdk.org/browse/JDK-8349727): [PPC] C1: Improve Class.isInstance intrinsic (**Enhancement** - P4)


### Reviewers
 * [Richard Reingruber](https://openjdk.org/census#rrich) (@reinrich - **Reviewer**)
 * [Varada M](https://openjdk.org/census#varadam) (@varada1110 - Committer)

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/23602/head:pull/23602` \
`$ git checkout pull/23602`

Update a local copy of the PR: \
`$ git checkout pull/23602` \
`$ git pull https://git.openjdk.org/jdk.git pull/23602/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 23602`

View PR using the GUI difftool: \
`$ git pr show -t 23602`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/23602.diff">https://git.openjdk.org/jdk/pull/23602.diff</a>

</details>
<details><summary>Using Webrev</summary>

[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/23602#issuecomment-2654844556)
</details>