@theRealAph 2025-02-19
Is this change actually worthwhile on PPC?