@TheRealMDoerr 2025-02-19
> Is this change actually worthwhile on PPC?

It's not a big gain. I've rather implemented it for parity with the other platforms. Maybe we want to remove the C++ version at some point of time?