@theRealAph 2025-02-20
> > Is this change actually worthwhile on PPC?
> 
> It's not a big gain. I've rather implemented it for parity with the other platforms. Maybe we want to remove the C++ version at some point of time?

Maybe, but I see no compelling reason to do so.