@TheRealMDoerr 2025-02-20
Thanks for the reviews and comments! It's a bit faster and does no harm. At least PPC64 is no longer in the way if we want to remove the C++ implementation.
/integrate